Prediction for double-constrained correspondence analysis (dc-CA)

Prediction of traits from environment, environment from traits and response from trait and environment data.

With type = "traits" and newdata = NULL, predict gives the fitted mean traits, i.e. the fitted community weighted means. With type = "env" and newdata = NULL, predict gives the fitted mean environment, i.e. the fitted species niche centroids.

Usage

# S3 method for class 'dcca'
predict(
  object,
  ...,
  type = c("env", "traits", "response", "reg_env", "reg_traits"),
  rank = "full",
  newdata = NULL
)

Arguments

object

return value of dc_CA.

...

Other arguments passed to the function (currently ignored).

type

type of prediction, c("env", "traits", "response", "reg_env", "reg_trait") for environmental values, values of traits, response (expected abundance), regression coefficients for environmental and trait predictors.

rank

rank or number of axes to use. Default "full" for all axes (no rank-reduction).

newdata

Data in which to look for variables with which to predict. For type = "reg_env" or "reg_trait" newdata is ignored. For type = "env" or "trait", newdata is a data frame of trait and environmental values, respectively, which are used for the prediction. If omitted, fitted values are generated. For type = "response", newdata is a list of two data frames with trait and environmental values in this order, e.g. list(traits = dataTraits, env = dataEnv).

Value

a matrix with the predictions. The exact content of the matrix depends on the type of predictions that are being made.

Details

Variables that are in the model but not in newdata are set to their weighted means in the training data. Predictions are thus at the (weighted) mean of the quantitative variables not included. Predictions with not-included factors are at the weighted mean (none of the factor effects are included).

For type = "response" and non-null newdata, the species weights of the training are used; the site weights are taken equal. Many of the predicted values may be negative, indicating expected absences (0) or small expected response values.

Regression coefficients obtained with type = "reg_env" or type = "reg_traits" are for standardized traits and environmental variables.

Examples

data("dune_trait_env")

# rownames are carried forward in results
rownames(dune_trait_env$comm) <- dune_trait_env$comm$Sites

mod <- dc_CA(formulaEnv = ~ A1 + Moist + Mag + Use + Condition(Manure),
             formulaTraits = ~ SLA + Height + LDMC + Condition(Seedmass) + Lifespan,
             response = dune_trait_env$comm[, -1],  # must delete "Sites"
             dataEnv = dune_trait_env$envir,
             dataTraits = dune_trait_env$traits)
#> Step 1: the CCA ordination of the transposed matrix with trait constraints,
#>         useful in itself and also yielding CWMs of the orthonormalized traits for step 2.
#> Call: cca(formula = tY ~ SLA + Height + LDMC + Condition(Seedmass) +
#> Lifespan, data = dataTraits)
#> 
#> -- Model Summary --
#> 
#>               Inertia Proportion Rank
#> Total         2.34899    1.00000     
#> Conditional   0.16671    0.07097    1
#> Constrained   0.51091    0.21750    4
#> Unconstrained 1.67137    0.71153   19
#> 
#> Inertia is scaled Chi-square
#> 
#> -- Eigenvalues --
#> 
#> Eigenvalues for constrained axes:
#>    CCA1    CCA2    CCA3    CCA4 
#> 0.21499 0.18622 0.07354 0.03616 
#> 
#> Eigenvalues for unconstrained axes:
#>    CA1    CA2    CA3    CA4    CA5    CA6    CA7    CA8 
#> 0.4386 0.2396 0.1938 0.1750 0.1429 0.1112 0.0854 0.0601 
#> (Showing 8 of 19 unconstrained eigenvalues)
#> 
#> Step 2: the RDA ordination of CWMs of the orthonormalized traits 
#>         of step 1 with environmental constraints:
#> Call: rda(formula = out1$CWMs_orthonormal_traits ~ A1 + Moist + Mag + Use +
#> Condition(Manure), data = out1$data$dataEnv)
#> 
#> -- Model Summary --
#> 
#>               Inertia Proportion Rank
#> Total          0.5109     1.0000     
#> Conditional    0.1297     0.2538    1
#> Constrained    0.1875     0.3669    4
#> Unconstrained  0.1938     0.3793    4
#> 
#> Inertia is variance
#> 
#> -- Eigenvalues --
#> 
#> Eigenvalues for constrained axes:
#>    RDA1    RDA2    RDA3    RDA4 
#> 0.08709 0.06049 0.02234 0.01755 
#> 
#> Eigenvalues for unconstrained axes:
#>     PC1     PC2     PC3     PC4 
#> 0.11386 0.04942 0.01888 0.01162 
#> 
#> mean, sd, VIF and canonical coefficients with their optimistic [!] t-values:
#>         Avg    SDS    VIF   Regr1   tval1
#> Manure 1.75 6.4614 9.3847 -0.2575 -1.0000
#> A1     4.85 9.4989 1.6967 -0.1167 -1.0660
#> Moist  2.90 7.8613 1.7658 -0.2097 -1.8779
#> MagBF  0.15 1.5969 4.5016 -0.1050 -0.5889
#> MagHF  0.25 1.9365 2.6715 -0.1489 -1.0841
#> MagNM  0.30 2.0494 9.5666 -0.2628 -1.0109
#> Use    1.90 3.4351 1.7825 -0.0537 -0.4782
#>                        Avg     SDS    VIF   Regr1   tval1
#> Seedmass            0.6543  0.6688 1.0784  0.1752  0.4465
#> SLA                24.6468  6.3438 1.1888  0.6785  1.6471
#> Height             25.1272 15.6848 1.3033 -0.5549 -1.2867
#> LDMC              244.5084 70.9729 1.1791  0.5597  1.3644
#> Lifespanperennial   0.9607  0.1944 1.0964 -0.3820 -0.9656
#> 
#>                weighted variance
#> total                      2.349
#> conditionT                 0.167
#> traits_explain             0.511
#> conditionE                 0.130
#> constraintsTE              0.187
#> attr(,"meaning")
#>                meaning                                                              
#> total          "total inertia"                                                      
#> conditionT     "inertia of the trait condition"                                     
#> traits_explain "trait-constrained inertia"                                          
#> conditionE     "trait-constrained inertia explained by the condition in formulaEnv" 
#> constraintsTE  "trait-constrained inertia explained by the predictors in formulaEnv"

# regression coefficients
predict(mod, type = "reg_env")
#>                SLA     Height        LDMC Lifespanannual Lifespanperennial
#> Manure -0.23423414  0.2540139  0.32912396    -0.37130684        0.37130684
#> A1      0.05226663  0.1025887 -0.08560552    -0.12390838        0.12390838
#> Moist  -0.18017363  0.1836804  0.04051036    -0.02158823        0.02158823
#> MagBF  -0.01504873  0.1005833  0.10645047    -0.25809310        0.25809310
#> MagHF  -0.05899194  0.1044629  0.03221358    -0.18057689        0.18057689
#> MagNM  -0.29468992  0.1739784  0.33095579    -0.39679946        0.39679946
#> Use    -0.04563582 -0.1100948 -0.17509044     0.03719925       -0.03719925
predict(mod, type = "reg_traits")
#>                            A1       Moist       MagSF        MagBF        MagHF
#> Seedmass          -0.02240045 -0.04978638 -0.01023869  0.041848273  0.009631558
#> SLA               -0.06845546 -0.16774065 -0.02649254  0.146106902  0.034893995
#> Height             0.15457078  0.15961290  0.02806485 -0.073452767  0.041610702
#> LDMC              -0.17828444 -0.10419687  0.01403797  0.005732252 -0.079309518
#> Lifespanperennial  0.08878173  0.02537622 -0.05136037 -0.005425793  0.007615053
#>                         MagNM          Use
#> Seedmass          -0.03147019  0.004500904
#> SLA               -0.12032477 -0.090142614
#> Height            -0.01014934  0.018558815
#> LDMC               0.05643595 -0.180870869
#> Lifespanperennial  0.04839256  0.080107192

# fit the mean traits at each site (20x6),
# that is CWM at each site
pred.traits <- predict(mod, type = "traits")
head(pred.traits)
#>        SLA   Height     LDMC Lifespanannual Lifespanperennial
#> 1 24.69054 23.04477 262.1525    0.047500047         0.9525000
#> 2 26.58938 22.47821 248.9740   -0.001033559         1.0010336
#> 3 24.27447 25.82005 259.4974    0.028101542         0.9718985
#> 4 24.25886 25.74429 259.7835    0.029235611         0.9707644
#> 5 26.80648 26.21316 241.2643    0.017241993         0.9827580
#> 6 26.11737 22.44992 230.8071    0.049338100         0.9506619

# fit the mean environment for each species (28x8)
# that is SNC of each species
pred.env <- predict(mod, type = "env")
head(pred.env)
#>                              A1    Moist     MagSF     MagBF     MagHF
#> Achillea millefolium   5.342999 3.326274 0.2987008 0.1012023 0.2658685
#> Agrostis stolonifera   4.529007 2.576826 0.2877323 0.1864249 0.2361958
#> Aira praecox           3.642275 2.771151 0.4244468 0.1056802 0.1759198
#> Alopecurus geniculatus 4.774155 2.531686 0.2733250 0.2228758 0.2796069
#> Anthoxanthum odoratum  3.917107 2.475008 0.3079576 0.1444949 0.1555197
#> Bellis perennis        4.651703 2.424855 0.2610354 0.2306304 0.2703286
#>                            MagNM      Use
#> Achillea millefolium   0.3342285 2.102427
#> Agrostis stolonifera   0.2896470 1.794929
#> Aira praecox           0.2939532 1.624523
#> Alopecurus geniculatus 0.2241924 1.874519
#> Anthoxanthum odoratum  0.3920278 1.618153
#> Bellis perennis        0.2380056 1.944774

pred.resp <- predict(mod, type = "response")
# pred has negative values and dc_CA cannot have negatives in the response
# so, modify pred.resp,
#whichgives about similar eigenvalues as the original data
pred.resp[pred.resp < 0] <- 0
mod3 <- dc_CA(formulaEnv = mod$formulaEnv,
              formulaTraits = mod$formulaTraits,
              response = pred.resp, 
              dataEnv = dune_trait_env$envir,
              dataTraits = dune_trait_env$traits)
#> Step 1: the CCA ordination of the transposed matrix with trait constraints,
#>         useful in itself and also yielding CWMs of the orthonormalized traits for step 2.
#> Call: cca(formula = tY ~ SLA + Height + LDMC + Condition(Seedmass) +
#> Lifespan, data = dataTraits)
#> 
#> -- Model Summary --
#> 
#>                 Inertia Proportion Rank
#> Total         1.661e-01  1.000e+00     
#> Conditional   8.889e-06  5.352e-05    1
#> Constrained   1.641e-01  9.881e-01    4
#> Unconstrained 1.971e-03  1.187e-02   10
#> 
#> Inertia is scaled Chi-square
#> 
#> -- Eigenvalues --
#> 
#> Eigenvalues for constrained axes:
#>    CCA1    CCA2    CCA3    CCA4 
#> 0.07673 0.05423 0.01895 0.01419 
#> 
#> Eigenvalues for unconstrained axes:
#>       CA1       CA2       CA3       CA4       CA5       CA6       CA7       CA8 
#> 0.0009360 0.0005522 0.0002421 0.0001346 0.0000649 0.0000307 0.0000100 0.0000003 
#>       CA9      CA10 
#> 0.0000001 0.0000000 
#> 
#> Step 2: the RDA ordination of CWMs of the orthonormalized traits 
#>         of step 1 with environmental constraints:
#> Call: rda(formula = out1$CWMs_orthonormal_traits ~ A1 + Moist + Mag + Use +
#> Condition(Manure), data = out1$data$dataEnv)
#> 
#> -- Model Summary --
#> 
#>                 Inertia Proportion Rank
#> Total         1.641e-01  1.000e+00     
#> Conditional   9.073e-05  5.529e-04    1
#> Constrained   1.635e-01  9.961e-01    4
#> Unconstrained 5.461e-04  3.328e-03    4
#> 
#> Inertia is variance
#> 
#> -- Eigenvalues --
#> 
#> Eigenvalues for constrained axes:
#>    RDA1    RDA2    RDA3    RDA4 
#> 0.07667 0.05405 0.01874 0.01399 
#> 
#> Eigenvalues for unconstrained axes:
#>       PC1       PC2       PC3       PC4 
#> 2.729e-04 2.527e-04 1.775e-05 2.780e-06 
#> 
#> mean, sd, VIF and canonical coefficients with their optimistic [!] t-values:
#>         Avg    SDS    VIF  Regr1   tval1
#> Manure 1.75 6.4614 9.3847 0.1783 26.9699
#> A1     4.85 9.4989 1.6967 0.0918 32.6413
#> Moist  2.90 7.8613 1.7658 0.1966 68.5625
#> MagBF  0.15 1.5969 4.5016 0.0637 13.9152
#> MagHF  0.25 1.9365 2.6715 0.1152 32.6500
#> MagNM  0.30 2.0494 9.5666 0.1860 27.8637
#> Use    1.90 3.4351 1.7825 0.0741 25.7072
#>                        Avg     SDS    VIF   Regr1    tval1
#> Seedmass            0.6554  0.6697 1.0755 -0.1577  -7.8655
#> SLA                24.6051  6.3645 1.1910 -0.6932 -32.8545
#> Height             25.1952 15.7708 1.3041  0.5404  24.4755
#> LDMC              244.6988 71.3268 1.1756 -0.6101 -29.1066
#> Lifespanperennial   0.9584  0.1997 1.0991  0.2877  14.1958
#> 
#>                weighted variance
#> total                      0.166
#> conditionT                 0.000
#> traits_explain             0.164
#> conditionE                 0.000
#> constraintsTE              0.163
#> attr(,"meaning")
#>                meaning                                                              
#> total          "total inertia"                                                      
#> conditionT     "inertia of the trait condition"                                     
#> traits_explain "trait-constrained inertia"                                          
#> conditionE     "trait-constrained inertia explained by the condition in formulaEnv" 
#> constraintsTE  "trait-constrained inertia explained by the predictors in formulaEnv"
mod3$eigenvalues / mod$eigenvalues
#>     dcCA1     dcCA2     dcCA3     dcCA4 
#> 0.8803851 0.8934459 0.8390431 0.7971850